scripts/requirements.txt
biopython
pyfaidx
gffutils
requests